ETL_Pipeline-main/requirements.txt
boto3>=1.34.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
